+++ R/xrd.find.Peak.R
@@ -1,5 +1,6 @@
-#' Finds a peak position based on a Gaussian fit with basic
-#' starting guess
+#' Finds Peak Position from Gaussian
+#'
+#' Fits a Gaussian form to a peak
 #'
 #' @param TwoTheta angle
 #' @param Intensity intensity signal
@@ -7,11 +8,15 @@
 #' @param Try.Sigma vector with peak widths used to start fitting
 #' @return fit peak position
 #' @examples
+#' library(stringr)
 #' filename = system.file("extdata", "2Theta.asc", package='rigakuXRD')
 #' d = xrd.read.ASC(filename)
-#' xrd.find.Peak(d$theta, d$I, 38.2)
+#' plot(d$theta,d$I,log='y',col='red')
+#' peak.pos = xrd.find.Peak(d$theta, d$I, 38.2)
+#' abline(v=peak.pos,col='blue')
 #'
 #' @import stats
+#' @import stringr
 #' @export
 xrd.find.Peak <- function(TwoTheta, Intensity, PeakPos,
                           Try.Sigma = c(0.1,0.2,0.15)) {