+++ R/xrd.get.DebyeScherrer.R
@@ -6,8 +6,7 @@
 #' @examples
 #' filename = system.file("extdata", "2Theta.asc", package='rigakuXRD')
 #' d = xrd.read.ASC(filename)
-#' q = xrd.get.PeakStats(d$theta, d$I, 38.2)
-#' xrd.get.PeakProminence(q)
+#' xrd.get.DebyeScherrer(d$theta, d$I, 38.2)
 #'
 #' @export
 xrd.get.DebyeScherrer <- function(PeakStats,