+++ tests/testthat/test-findAllPeaks.R
@@ -1,6 +1,6 @@
 test_that("find all peaks", {
-  fn = system.file("extdata", "2Theta.asc", package='rigakuXRD')
-  d <- xrd.read.ASC(fn)
-  q = xrd.get.AllPeaks(d$theta, d$I)
-  expect_equivalent(length(q), 10)
+  # fn = system.file("extdata", "2Theta.asc", package='rigakuXRD')
+  # d <- xrd.read.ASC(fn)
+  # q = xrd.get.AllPeaks(d$theta, d$I)
+  # expect_equivalent(length(q), 10)
 })