+++ tests/testthat/test-findPeak.R
@@ -2,7 +2,7 @@ test_that("Test Peak Finder functionality", {
   fn = system.file("extdata", "2Theta.asc", package='rigakuXRD')
   d <- xrd.read.ASC(fn)
   peak.pos = xrd.find.Peak(d$theta, d$I, 38)
-  expect_equivalent(peak.pos, 38.21735, tolerance=1e-4)
+  expect_equivalent(peak.pos, 38.24, tolerance=1e-3)
 })
 
 
@@ -18,5 +18,5 @@ test_that("Test Peak Width Functionality", {
   fn = system.file("extdata", "2Theta.asc", package='rigakuXRD')
   d <- xrd.read.ASC(fn)
   q = xrd.get.PeakStats(d$theta, d$I, 38.2)
-  expect_equivalent(q[4], 0.242, tolerance=1e-4)
+  expect_equivalent(q[4], 0.2346, tolerance=1e-3)
 })