+++ tests/testthat/test-rigakuXRD.R
@@ -1,8 +1,0 @@
-library(testthat)
-
-test_that("ASC test file exists", {
-  fn = system.file("extdata", "2Theta.asc", package='rigakuXRD')
-  expect_true(
-    file.exists(fn)
-  )
-})