+++ NEWS.md
@@ -1,3 +1,4 @@
 # rigakuXRD 0.2.0
 
+* use `xrd.importData` to read XRD data files
 * Added a `NEWS.md` file to track changes to the package.